move cli interface to separate module
