+++ busy_beaver/apps/events/blueprint.py
@@ -2,5 +2,5 @@ from flask import blueprints
 
 events_bp = blueprints.Blueprint("events", __name__)
 
-from . import workflow  # noqa isort:skip
+from . import cli  # noqa isort:skip
 from .event_database import task  # noqa isort:skip
